requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
joblib==1.4.2
pillow==10.4.0
numpy==1.26.4
pandas==2.2.3
tensorflow-metal==1.2.0
tensorflow==2.16.2
scikit-learn==1.5.2
python-multipart